Update for adobe edge Users
This is my first time using Github - this code did not work with my program, because I am using adobe edge code. I had to change this line of code to make it work. I am not sure how to implement it properly into an IF statement, so I just commented it out. I apologize If I am doing something wrong here.